webui-selenium-test_withsleep.py replaces webui-selenium-test.py since that is the file I'm working on.
